requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
pymongo==4.6.1
dnspython==2.4.2
itsdangerous==2.1.2
cloudinary==1.36.0
pyzbar==0.1.9
Pillow==10.0.1
pdf2image==1.17.0
requests==2.31.0
nltk==3.8.1
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
python-dotenv==1.0.0
gunicorn==21.2.0
